♻️ rename utils to envUtils
